SILGen: Split Objective-C specific bits out of test/SILGen/vtables.swift
